fix(low-router): improve tests
